requirements.txt
django
psycopg2-binary
python-dotenv
django-extensions
djangorestframework
djangorestframework-simplejwt
drf-yasg
